Update news.vue to include filtering by tags (#54)
* Update news.vue to include filtering by tags

Enabled filtering of articles so that when a tag is clicked only articles with that particular tag are shown. This can be done for multiple tags at once. If a selected tag is clicked again, it is unselected. A reset button was also added to clear all filtering.

* Updated news.vue

Made changes to code and page appearance based on input from Kevin.