Add Recent News table to home page (#60)
* Add Recent News table to home page

Added a table with recent news on the home page, with a link to the news page.

* Update index.vue

Changed anchor tag to nuxt-link tag on More.../Plus... hyperlink

* Update index.vue

Updated the news section to use an unordered list instead of the table format. Reformatted the html in the news section.

* Formatting changes to news.vue and index.vue

Inserted v-html property into news.vue to properly show the html formatting in the news blurbs. Fixed the margins of the news articles on the home page to align with the header.

* Update index.vue

Changed span tag to div tag