Edit data/search.vue
Changed query section in data/search.vue to load all the results from the ES query made. After the first 500 results are loaded to the page, the remaining results are loaded in queries producing 500 results. This is time-consuming but I am finding queries for larger datasets are timing out. I am currently looking for alternate methods.