Edit country dropdown list in Data Search / Download page
Edited country dropdown list in data/search.vue so that regardless of how zoomed in to the map you are, the dropdown list of countries always includes all countries, not just the countries visible on the map. The stations list, however, is narrowed to only the applicable stations once a country is selected.
(additionally, some formatting changes  to pass lintfix)