Server-side sorting of Data Search Table (#89)
* Data Search Table Edit

Data Search Table edits to column headers and entries.

* Add hyperlinks to PeerDataRecords Search results table

Add GAW ID and Contributor URLs to PeerDataRecords Data Search results table

* Server-side sorting of Data Search Table

Enabled server-side sorting of the results table columns on the Data Search & Download page. Sorting by properties of type DATE_FORMAT is not presently enabled, so a change must be made to the API to enable sorting for the associated columns.