Bounding box filtering edit on Data Search page (#93)
* Add Longitude and Latitude tags to metrics chart

Add Longitude and Latitude tags to metrics chart on Data Seach page. These tags show the limits of the bounding box used in the data search.

* Make Data Search Filtering by Bounding Box Optional

Made filtering on the Data Search page by the map's bounding box optional. Metrics chart does not update when the box is moved if bbox filtering is disabled.

* Disable bounding box filter when a country or station is selected

Disable bounding box filter when a country or station is selected

* Edits to Bounding Box Filtering

Combined Latitude/Longitude tags, changed text on the switch and switch position.

* Positioning edits to selectable-map and bbox switch

Positioning/margin edits to selectable-map and bbox switch